Health check shield API
Implementation of an API which temporarly prevents the task from being killed when it's unhealthy.
This is required to support troubleshooting scenarios when the app is "legally" unhealthy, for example when taking a memory dump.

API:
- GET /v2/shield: show current shields
- PUT /v2/shield/<task_id>?ttl=<ttl>: add the shield
- DELETE /v2/shield/<task_id>: remove the shield

Regarding authorization:
- read access (GET) only require authentication
- write access (PUT & DELETE) requires UpdateRunSpec of specified task
If there's no application delete is allowed (to be able to cleanup shields of stale apps using API)