Let TaskReplaceActor ignore outdated health checks
We observed issues when mixing Marathon and Mesos health checks.
Sometimes, the Mesos health checks get duplicated, and one 'replica'
keeps saying the instance is not alive, even if it is.
These broken/outdated health checks can be spotted easily, since their last
update timestamp is prior to the timestamp at which the task was
staged at.
Ignoring these outdated health checks allows the TaskReplaceActor to
correctly kill the healthy instances, and perform the deployment of the
new version of the application.

JIRA: MESOS-4437