Let HealthCheckActor forget outdated healthcheck result
We observed an issue with Mesos healthchecks.
When a task failed its Mesos HC, it was killed and replaced by a new
task - this was expected. What was unexpected however was that the new
task was also seen as unhealthy (even if it was running correctly), with
2 healthcheck results attached to it (an outdated one reporting a
failure, and a correct one reporting success).

This commit fixes the issue by letting the HealthCheckActor forget
outdated healthckeck result.

JIRA: MESOS-4437