Recycle unhealthy instances when multiple app versions run concurrently
Anti-snowball is implemented inside the HealthCheckActor. A
HealthCheckActor keeps track of health check results for a given version
of a given app. Since several versions of an app can run concurrently,
sometimes the HealthCheckActor cannot take the right decision because it
has only a partial view of the application state.
This patch allows the HealthCheckActor to have a global view of the
application healthiness by communicating with the AppHealthCheckActor,
when it is needed.

Co-authored-by: Julien DOCHE <j.doche@criteo.com>